+++ Changes.md
@@ -5,6 +5,7 @@ to add your changes here when sending pull requests. Also send corrections if
 you spot any mistakes.
 
 ## Unreleased
+* Adding __dirname before the directory not longer necessary
 * Added name and path to resolve function
 
 ## v2.0.0 (2015-10-17)