+++ Changes.md
@@ -4,7 +4,7 @@ This file is a manually maintained list of changes for each release. Feel free
 to add your changes here when sending pull requests. Also send corrections if
 you spot any mistakes.
 
-## HEAD
+## v2.1.0 (2016-12-09)
 
 * Accept a function for "filter" option #27 #31